requirements.txt

Pillow
lunar-python